feat(explore): Adds prebuilt queries to Explore (#90720)
Adds prebuilt queries to Explore:
- `PREBUILT_SAVED_QUERIES` are saved as hardcoded dicts in
`explore_saved_queries.py`.

- On the first time `GET` is called on `ExploreSavedQueriesEndpoint`, we
insert `PREBUILT_SAVED_QUERIES` entries into the saved queries db table,
if they do not already exist. We also update or delete saved prebuilt
queries in the db if there are changes detected (see
`sync_prebuilt_queries`).

- If it's the first time a user is fetching prebuilt queries, we also
"star" them for the user by default. See
`sync_prebuilt_queries_starred`.

- This allows users to (mostly) interact with prebuilt queries in the
same way as user generated saved queries (ie star, reorder star
position, search filter, search sort, duplicate, etc). This also allows
Sentry to add, change, or delete any prebuilt queries in the future. We
do not allow users to delete or edit prebuilt queries.

+++ src/sentry/explore/endpoints/explore_saved_query_detail.py
@@ -104,6 +104,9 @@ class ExploreSavedQueryDetailEndpoint(ExploreSavedQueryBase):
 
         self.check_object_permissions(request, query)
 
+        if query.prebuilt_id is not None:
+            return self.respond(status=400, message="Cannot modify prebuilt queries")
+
         try:
             params = self.get_filter_params(
                 request, organization, project_ids=request.data.get("projects")
@@ -147,6 +150,9 @@ class ExploreSavedQueryDetailEndpoint(ExploreSavedQueryBase):
             return self.respond(status=404)
 
         self.check_object_permissions(request, query)
+
+        if query.prebuilt_id is not None:
+            return self.respond(status=400, message="Cannot delete prebuilt queries")
 
         query.delete()
